backend/requirements.txt
supabase==2.3.1
python-dotenv==1.0.0
elevenlabs==0.2.9
python-multipart==0.0.6
fastapi==0.95.2
uvicorn==0.22.0